+++ CHANGELOG.md
@@ -8,3 +8,6 @@
 - added baseline code to signal.py that tunes to 315 MHz, recieves samples, and calculates their avg amplitude
 ### Added 
 - dsp_demodulation.py
+
+### Added
+- created threshold for signal burst in signal.py